--- a/.gitignore
+++ b/.gitignore
@@ -25,4 +25,4 @@ _testmain.go
 
 coverage.out
 
-benchmark_repos/
+benchmark/