A faster implementation of `GetCommitsInfo`, addresses https://github.com/go-gitea/gitea/issues/491 and https://github.com/go-gitea/gitea/issues/502.

The previous implementation made a call to `git log` for each entry, each of which required scanning through the commit history. This new implementation instead makes a single call to `git log`. This is faster, because it involves scanning the commit history only once.

__BENCHMARK RESULTS__:
Shoutout to @sapk for the benchmark tests he wrote (#54), which I have stolen here.

Old implementation:
```
BenchmarkEntries_GetCommitsInfo/gitea-4         	     200	  70670229 ns/op
BenchmarkEntries_GetCommitsInfo/manyfiles-4     	       1	35907609610 ns/op
BenchmarkEntries_GetCommitsInfo/moby-4          	      50	 204585469 ns/op
BenchmarkEntries_GetCommitsInfo/go-4            	      20	 640497304 ns/op
BenchmarkEntries_GetCommitsInfo/linux-4         	      20	 829981474 ns/op
```

New implementation:
```
BenchmarkEntries_GetCommitsInfo/gitea-4         	     500	  34284443 ns/op
BenchmarkEntries_GetCommitsInfo/manyfiles-4     	      30	 424753714 ns/op
BenchmarkEntries_GetCommitsInfo/moby-4          	     100	 109430547 ns/op
BenchmarkEntries_GetCommitsInfo/go-4            	      30	 521336982 ns/op
BenchmarkEntries_GetCommitsInfo/linux-4         	      20	 813171326 ns/op
```

__IMPLEMENTATION DETAILS__:
Gets the 16 latest commits affecting the relevant entries (`git log --name-only -16 HEAD -- <treePath>`). The output of this command containing which files were affect by each commit. Scan through this list of cmmit, and stop once a commit has been found for each entry.

If you go through the first batch of 16 commits, you get then next 32 commits (`git log --name-only -32 <last-commit-from-first-batch>^ -- entry1 entry2 ...`); each time you double the number of commits. This ensures that in the common case you'll only have to read in a small number (16) of commits, but you also won't have to make too many calls to `git log` if you need to go further into the commit history.

Finally, if you are looking for 32 or fewer entries, manually list out each entry in the `git log` command (`git log --name-only -- entry1 entry2...`) to support a more targeted search.