@ethantkoenig 2017-05-23
Trace lines removed, sorry about that