@lunny 2017-05-23
Are there some performance tests?