@sapk 2017-05-24
We could add golang benchmark or more simple benchmark like a script in a folder contrib ?