@sapk 2017-05-24
Maybe I have made a mistake in my benchmark (#54) but I don't find optimization on some big repo. 

Before PR : 
![before PR](https://cloud.githubusercontent.com/assets/4052400/26405911/f7510ee8-4096-11e7-81c4-b4cbf2f906a9.png)

After PR : 
![after PR](https://cloud.githubusercontent.com/assets/4052400/26405853/d1917058-4096-11e7-914f-120dd983bf69.png)