@ethantkoenig 2017-05-24
@sapk Thanks for writing the tests, I'll look into it