@ethantkoenig 2017-05-24
If you try a repo with many files (e.g. https://github.com/ethantkoenig/manyfiles), you should see a noticeable speed-up (old: 50 seconds, new: 5 seconds on my laptop).

I'll try to look for how to make my implementation comparable to the old implementation for non-pathological cases.