@ethantkoenig 2017-05-25
@sapk I've found a faster implementation that improves all 5 benchmarks test (some slightly, others quite significantly). Let me know what you think