@sapk 2017-05-25
Globally, I have not review implementation (yet), benchmark show some little improvement that would be good to have. I made some comments on the benchmark part. I would have prefer that you cherry-pick my commits/PR and made change after but it could pass for this PR ^^.