@sapk 2017-05-26
@ethantkoenig couldn't this have some concurrency like old code in order to speed up again ? I know that more routine is not always the solution and that maybe you already test that.