@ethantkoenig 2017-05-26
@sapk I don't think this implementation is as amenable to concurrency as the old one. As far as I can tell, the main benefit from using concurrency in the old implementation was to make calls to `git log ...` in parallel. The new implementation makes far fewer calls to `git log ...`, and these calls must occur in succession (each one depends on the result of the previous one), so I don't think we'd see much gain from introducing concurrency.