@sapk 2017-05-26
@ethantkoenig looking at it could have some concurrency but with a different format/strategy.  For example a controller starting routines executing git log history and returning by chan the path matching. The controller stop starting new routine when all path are completed. 

This could be done later. This PR give an improvement and LGTM. (except panic() that need to be changed)
This PR could have a test to check regression but I haven't a good idea how to do that so maybe another time. ^^